% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_best_group.R
\name{assign_best_group}
\alias{assign_best_group}
\title{Rattacher chaque commune candidate au groupe voisin le plus proche}
\usage{
assign_best_group(data, group_var, vars_attr, profiles)
}
\arguments{
\item{data}{Un objet sf avec colonne "candidate" et "id_row"}

\item{group_var}{Variable de groupe}

\item{vars_attr}{Variables attributaires}

\item{profiles}{Profils moyens par groupe (depuis compute_profiles)}
}
\value{
data.frame avec id_row et best_group
}
\description{
Rattacher chaque commune candidate au groupe voisin le plus proche
}
