% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_neighbors.R
\name{build_neighbors}
\alias{build_neighbors}
\title{Construire la matrice de voisinage spatial}
\usage{
build_neighbors(data, type = "queen")
}
\arguments{
\item{data}{Un objet sf}

\item{type}{"queen" ou "rook"}
}
\value{
Un objet nb (spdep)
}
\description{
Construire la matrice de voisinage spatial
}
