% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_affinity.R
\name{compute_affinity}
\alias{compute_affinity}
\title{Calculer les affinités attributaires des communes frontalières}
\usage{
compute_affinity(data, nb, group_var, vars_attr)
}
\arguments{
\item{data}{Un objet sf}

\item{nb}{Matrice de voisinage (spdep)}

\item{group_var}{Nom de la variable de groupe (ex: "EPCI")}

\item{vars_attr}{Vecteur de noms de variables attributaires}
}
\value{
Un data.frame avec dist_intra, dist_extra, affinite
}
\description{
Calculer les affinités attributaires des communes frontalières
}
