% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_profiles.R
\name{compute_profiles}
\alias{compute_profiles}
\title{Calculer les profils moyens par groupe}
\usage{
compute_profiles(data, group_var, vars_attr)
}
\arguments{
\item{data}{Un objet sf}

\item{group_var}{Variable de groupe}

\item{vars_attr}{Variables attributaires}
}
\value{
data.frame des profils moyens par groupe
}
\description{
Calculer les profils moyens par groupe
}
