% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_isolates.R
\name{remove_isolates}
\alias{remove_isolates}
\title{Reintegrer les communes candidates isolees}
\usage{
remove_isolates(data, nb)
}
\arguments{
\item{data}{Un objet sf avec colonne "candidate"}

\item{nb}{Matrice de voisinage}
}
\value{
data avec candidats isoles remis a FALSE
}
\description{
Reintegrer les communes candidates isolees
}
