% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialRegroup.R
\name{spatialRegroup}
\alias{spatialRegroup}
\title{Regroupement spatial par affinite attributaire}
\usage{
spatialRegroup(
  data,
  group_var,
  vars_attr,
  remove_isolates = TRUE,
  nb_type = "queen",
  verbose = TRUE
)
}
\arguments{
\item{data}{sf object}

\item{group_var}{Variable de groupe initial (ex: "EPCI")}

\item{vars_attr}{Variables attributaires pour le calcul d'affinite}

\item{remove_isolates}{Reintegrer les isolats (defaut TRUE)}

\item{nb_type}{"queen" ou "rook"}

\item{verbose}{Messages de progression (defaut TRUE)}
}
\value{
An sf object identical to the input with two new columns: 'EPCI_regroup' (new group assignment) and 'candidate' (logical, TRUE if reclassified).
}
\description{
Regroupement spatial par affinite attributaire
}
